♻️  application_controllerに置いていたsession関連のメソッドをmoduleに切り出した
